Add module algorithm and hover
